+++ examples/using-redux/gatsby-browser.js
@@ -1,16 +1,2 @@
-import React from "react"
-import { Provider } from "react-redux"
-
-import createStore from "./src/state/createStore"
-
-const store = createStore()
-
-export const wrapRootComponent = ({ Root }) => {
-  const ConnectedRootComponent = () => (
-    <Provider store={store}>
-      <Root />
-    </Provider>
-  )
-
-  return ConnectedRootComponent
-}
+import wrapWithProvider from "./wrap-with-provider"
+export const wrapRootElement = wrapWithProvider